block slashing protection unit test

+++ slashing-protection/src/main/resources/migrations/postgresql/V1__initial.sql
@@ -1,17 +1,17 @@
 CREATE TABLE validators (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   public_key BYTEA NOT NULL,
   UNIQUE(public_key)
 );
 CREATE TABLE signed_blocks (
-    validator_id BIGINT NOT NULL,
+    validator_id INTEGER NOT NULL,
     slot NUMERIC(20) NOT NULL,
     signing_root BYTEA NOT NULL,
     FOREIGN KEY(validator_id) REFERENCES validators(id),
     UNIQUE (validator_id, slot)
 );
 CREATE TABLE signed_attestations (
-    validator_id BIGINT,
+    validator_id INTEGER,
     source_epoch NUMERIC(20) NOT NULL,
     target_epoch NUMERIC(20) NOT NULL,
     signing_root BYTEA NOT NULL,
